Extract rx2 test helpers
