+++ libraries/rib-base-test-rx2/build.gradle
@@ -1,0 +1,41 @@
+configureAndroidLibrary(project)
+
+android {
+    defaultConfig {
+        versionCode 1
+        versionName "1.0"
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+        }
+    }
+
+
+    lintOptions {
+        lintConfig file("lint.xml")
+    }
+}
+
+dependencies {
+    compileOnly deps.android.annotations
+
+    compileOnly project(':libraries:rib-base')
+    compileOnly project(':libraries:rib-base-test')
+    compileOnly project(':libraries:rib-base-test-activity')
+    compileOnly project(':libraries:rib-rx2')
+    compileOnly deps.external.rxrelay2
+    compileOnly deps.external.rxjava2
+
+    implementation deps.external.kotlinStdlib
+    implementation deps.android.annotations
+    implementation deps.android.runner
+    implementation deps.test.junit
+    implementation deps.test.mockito
+
+    api deps.android.rules
+}
+
+apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
