+++ libraries/rib-base-test-rx2/gradle.properties
@@ -1,0 +1,3 @@
+POM_NAME=RIBs (Base Test Rx2 helpers)
+POM_ARTIFACT_ID=rib-base-test-rx2
+POM_PACKAGING=android
