+++ libraries/rib-base-test-rx2/lint.xml
@@ -1,0 +1,6 @@
+<lint>
+    <issue id="InvalidPackage">
+        <ignore path="**/byte-buddy-agent-*.jar"/>
+        <ignore path="**/mockito-core-*.jar"/>
+    </issue>
+</lint>
