+++ libraries/rib-base-test-rx2/src/main/AndroidManifest.xml
@@ -1,0 +1,3 @@
+<manifest package="com.badoo.common.ribs">
+
+</manifest>
