+++ settings.gradle
@@ -12,6 +12,7 @@ include ':app-example'
 include ':libraries:rib-base'
 include ':libraries:rib-base-test'
 include ':libraries:rib-base-test-activity'
+include ':libraries:rib-base-test-rx2'
 include ':libraries:rib-debug-utils'
 include ':libraries:rib-portal'
 include ':libraries:rib-portal-rx'
